adds routing to tv-eats web app, adds other pages to be edited soon
