+++ package.json
@@ -9,9 +9,13 @@
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.96",
     "@types/react-dom": "^18.2.25",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1",
     "react-scripts": "^5.0.1",
+    "sort-by": "^1.2.0",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
