+++ src/index.tsx
@@ -1,19 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Home from './components/Home';
 import reportWebVitals from './reportWebVitals';
-import Navbar from './components/Navbar';
+import App from './App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    
-    <Navbar />
-    <Home />
-    
+    <App />
   </React.StrictMode>
 );
