+++ src/pages/Contact.tsx
@@ -1,0 +1,10 @@
+import Navbar from "../components/Navbar"
+
+export default function Contact() {
+    return (
+        <div>
+            <Navbar />
+            <h1>Contact</h1>
+        </div>
+    );
+}
