+++ src/pages/Error.tsx
@@ -1,0 +1,8 @@
+
+export default function Error() {
+    return (
+        <div>
+            <h1>Error 404: Page not found!</h1>
+        </div>
+    );
+}
