+++ src/pages/Restaurants.tsx
@@ -1,0 +1,10 @@
+import Navbar from "../components/Navbar"
+
+export default function Restaurants() {
+    return (
+        <div>
+            <Navbar />
+            <h1>Restaurants</h1>
+        </div>
+    );
+}
