+++ src/pages/Shows.tsx
@@ -1,0 +1,10 @@
+import Navbar from "../components/Navbar"
+
+export default function Shows() {
+    return (
+        <div>
+            <Navbar />
+            <h1>Shows</h1>
+        </div>
+    );
+}
